Restructure python module
