+++ wrappers/python/mypy.ini
@@ -1,0 +1,25 @@
+[mypy]
+files=solrstice
+warn_unused_configs=True
+
+# Dynamic typing
+disallow_any_unimported = True
+disallow_any_generics=True
+disallow_subclassing_any = True
+
+# Untyped definitions and calls
+disallow_untyped_calls = True
+disallow_untyped_defs = True
+disallow_incomplete_defs = True
+check_untyped_defs = True
+disallow_untyped_decorators = True
+
+# Warnings
+warn_redundant_casts = True
+## Conflict between mypy and pyright
+warn_unused_ignores = False
+warn_no_return = True
+
+# Misc
+allow_redefinition = True
+strict_equality = True
