+++ wrappers/python/solrstice/collection.pyi
@@ -1,13 +1,14 @@
 from typing import List, Optional
 
-from solrstice.hosts import SolrServerContext
+from solrstice import SolrServerContext
+
 
 async def create_collection(
-    context: SolrServerContext,
-    name: str,
-    config: str,
-    shards: Optional[int] = 1,
-    replication_factor: Optional[int] = 1,
+        context: SolrServerContext,
+        name: str,
+        config: str,
+        shards: Optional[int] = 1,
+        replication_factor: Optional[int] = 1,
 ) -> None:
     """
     Create a collection on the Solr server.
@@ -18,13 +19,14 @@ async def create_collection(
     :param shards: The number of shards to create.
     :param replication_factor: The replication factor to use.
     """
+
 
 def create_collection_blocking(
-    context: SolrServerContext,
-    name: str,
-    config: str,
-    shards: Optional[int] = 1,
-    replication_factor: Optional[int] = 1,
+        context: SolrServerContext,
+        name: str,
+        config: str,
+        shards: Optional[int] = 1,
+        replication_factor: Optional[int] = 1,
 ) -> None:
     """
     Create a collection on the Solr server.
@@ -35,6 +37,7 @@ def create_collection_blocking(
     :param shards: The number of shards to create.
     :param replication_factor: The replication factor to use.
     """
+
 
 async def get_collections(context: SolrServerContext) -> List[str]:
     """
@@ -44,6 +47,7 @@ async def get_collections(context: SolrServerContext) -> List[str]:
     :return: The list of collections on the Solr server.
     """
 
+
 def get_collections_blocking(context: SolrServerContext) -> List[str]:
     """
     Get the list of collections on the Solr server.
@@ -51,6 +55,7 @@ def get_collections_blocking(context: SolrServerContext) -> List[str]:
     :param context: The Solr server context.
     :return: The list of collections on the Solr server.
     """
+
 
 async def collection_exists(context: SolrServerContext, name: str) -> bool:
     """
@@ -61,6 +66,7 @@ async def collection_exists(context: SolrServerContext, name: str) -> bool:
     :return: True if the collection exists, False otherwise.
     """
 
+
 def collection_exists_blocking(context: SolrServerContext, name: str) -> bool:
     """
     Check if a collection exists on the Solr server.
@@ -70,6 +76,7 @@ def collection_exists_blocking(context: SolrServerContext, name: str) -> bool:
     :return: True if the collection exists, False otherwise.
     """
 
+
 async def delete_collection(context: SolrServerContext, name: str) -> None:
     """
     Delete a config from the Solr server.
@@ -77,6 +84,7 @@ async def delete_collection(context: SolrServerContext, name: str) -> None:
     :param context: The Solr server context.
     :param name: The name of the collection to delete.
     """
+
 
 def delete_collection_blocking(context: SolrServerContext, name: str) -> None:
     """
