+++ wrappers/python/tests/test_def_type.py
@@ -8,8 +8,8 @@ from helpers import (
     wait_for_solr,
 )
 
-from solrstice.def_type import DismaxQuery, EdismaxQuery, LuceneQuery
-from solrstice.queries import SelectQuery
+from solrstice import DismaxQuery, EdismaxQuery, LuceneQuery
+from solrstice import SelectQuery
 
 
 @pytest.fixture()
