+++ wrappers/python/tests/test_select.py
@@ -8,7 +8,7 @@ from helpers import (
     wait_for_solr,
 )
 
-from solrstice.queries import SelectQuery
+from solrstice import SelectQuery
 
 
 @pytest.fixture()
